Test unitário create and delete patient

+++ server/src/use-cases/patients/delete-patient-use-case.ts
@@ -10,6 +10,11 @@ export class DeletePatientUseCase {
   ){}
 
   async delete({id}: DeletePatientUseCaseRequest) {
+
+    if (!id) {
+      throw new Error('Id is required');
+    }
+
     await this.patientsRepository.delete({id});
   }
 }
